+++ packages/alibaba/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/alibaba
 
+## 2.0.0-canary.46
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+  - @ai-sdk/openai-compatible@3.0.0-canary.44
+
 ## 2.0.0-canary.45
 
 ### Patch Changes