+++ packages/otel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/otel
 
+## 1.0.0-canary.74
+
+### Patch Changes
+
+- ai@7.0.0-canary.128
+
 ## 1.0.0-canary.73
 
 ### Patch Changes